[quantization] Extracted quantize/dequantize out of linear.
